Revert "Refactor index forms to use a template"
