+++ .anvil_editor.yaml
@@ -3,7 +3,6 @@ unique_ids:
     theme.css: 6G44GIEQQRJLILWU6GQSQO5BTDFT5ZI3
   forms:
     forms.components.DetailActions: '1731068516535911935559171.9653'
-    forms.components.Index: '1740065989081166916053507.60074'
     forms.components.ReactiveTabulator: '1731084304741819530292745.9788'
     forms.layouts.Main: '1731062800708407524763448.735'
     forms.pages.author.Detail: '1731063504077749734619018.2582'
