--- client_code/forms/components/Index/form_template.yaml
@@ -1,23 +1,0 @@
-components_by_slot:
-  content:
-  - components:
-    - data_bindings:
-      - {code: self.title, property: text, writeback: false}
-      name: heading
-      properties: {}
-      type: form:dep_lin1x4oec0ytd:_Components.Heading
-    - event_bindings: {click: new_button_click}
-      layout_properties: {}
-      name: new_button
-      properties: {icon: 'mi:add', text: New}
-      type: form:dep_lin1x4oec0ytd:_Components.Button
-    layout_properties: {}
-    name: flow_panel_1
-    properties: {align: justify, vertical_align: middle}
-    type: FlowPanel
-  - layout_properties: {}
-    name: tabulator
-    properties: {}
-    type: form:forms.components.ReactiveTabulator
-is_package: true
-layout: {type: 'form:forms.layouts.Main'}
