+++ client_code/forms/pages/book/Index/form_template.yaml
@@ -1,4 +1,22 @@
 components_by_slot:
-  content: []
+  content:
+  - components:
+    - layout_properties: {}
+      name: heading_1
+      properties: {text: Books}
+      type: form:dep_lin1x4oec0ytd:_Components.Heading
+    - event_bindings: {click: new_button_click}
+      layout_properties: {}
+      name: new_button
+      properties: {icon: 'mi:add', text: New}
+      type: form:dep_lin1x4oec0ytd:_Components.Button
+    layout_properties: {}
+    name: flow_panel_1
+    properties: {align: justify, vertical_align: middle}
+    type: FlowPanel
+  - layout_properties: {}
+    name: tabulator
+    properties: {}
+    type: form:forms.components.ReactiveTabulator
 is_package: true
 layout: {type: 'form:forms.layouts.Main'}
